requirements.txt
requests==2.31.0
hikari==2.0.0.dev122
python-dotenv==1.0.1
black==22.3.0
pre-commit==3.6.2
pymongo==4.6.2
httpx==0.27.0
watchdog==4.0.0
motor==3.3.2
